Superuser: Translation fixes
R.string.disabled is an AOSP string present in Settings. It was
overriding the same string that was present in strings.xml of Superuser,
causing translations to be ineffective. Change the string in superuser
to R.string.access_disabled to avoid this conflict.

Patch Set 3: Rebase

Patch Set 4: Remove carriage returns and tabs from files.

Change-Id: I22e2cf51d6931fd79c39d70f57142120cab93a44